Fix 64Bit kernel stack corruption
The inode_init_security hook passes a pointer to in to a function that
expects a pointer to size_t.

Aegis-Change: anoubis.10.972.C874 Delta 460
TE BLE

+++ linux/linux-git/security/anoubis/anoubis_core.c
@@ -825,7 +825,7 @@ static int ac_inode_init_security(struct inode *inode, struct inode *dir,
 	unsigned int mymagic, lastmagic = 0;
 	char *myname = NULL, *name = NULL;
 	void *myvalue = NULL, *value = NULL;
-	int mylen = 0, len = 0;
+	size_t mylen = 0, len = 0;
 
 	if (original_ops->inode_init_security) {
 		ret = original_ops->inode_init_security(inode, dir,
